Rename Lilith to Avalon.
Avalon is the direct copy of Lilith v0.2.0 (this commit corresponds
to Lilith's bbd8e8147c2d71f175ced32855ca19f3607f0baa). The main reason
for renaming is difficulty of pronunciation and difference between
languages in the final "th" ([θ]) which makes it hard to speak about.
Since Julia package system doesn't support package renaming, the only
reliable way to create a new name is to create a new package. This is
how we got to where we are.